2021-11-14 redis组件更新

+++ dptms-common/dptms-common-security-starter/src/main/java/com/darren1112/dptms/common/security/starter/properties/SecurityProperties.java
@@ -38,10 +38,10 @@ public class SecurityProperties {
     /**
      * token有效期(s)
      */
-    private long accessTokenExpired = 60 * 15;
+    private int accessTokenExpired = 60 * 15;
 
     /**
      * 刷新token有效期(s)
      */
-    private long refreshTokenExpired = 60 * 60 * 24 * 7;
+    private int refreshTokenExpired = 60 * 60 * 24 * 7;
 }
